+ DS_BSTree + pack_train

+++ src/org/shu/job/Xiangjia.java
@@ -1,11 +1,9 @@
 package org.shu.job;
 
 import java.util.Scanner;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
 import java.math.BigDecimal;
-import java.util.Scanner;
+import java.math.BigInteger;
+
 /*
  * 设计一个算法能够完成两个字符串存储的整数进行想加操作，对非法输入则返回“error”
  * 输入为一行，包含两个字符串，逗号分隔，字符串的长度在[1,100]。输出为一行
@@ -23,9 +21,10 @@ public class Xiangjia {
 		boolean isNumTwo = s[1].matches("[0-9]+");
 
 		if (isNumOne && isNumTwo) {
-			BigDecimal one = new BigDecimal(s[0]);
-
-			BigDecimal two = new BigDecimal(s[1]);
+//			BigDecimal one = new BigDecimal(s[0]);
+			BigInteger one =new BigInteger(s[0]);
+//			BigDecimal two = new BigDecimal(s[1]);
+			BigInteger two=new BigInteger(s[1]);
 
 			System.out.println(one.add(two));
 		} else {
